--- backend/src/main/java/com/example/nbe341team02/domain/delivery/DeliveryTimePolicy.java
@@ -1,33 +1,0 @@
-package com.example.nbe341team02.domain.delivery;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.ColumnDefault;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Getter
-public class DeliveryTimePolicy {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, updatable = false)
-    private LocalTime deliveryTime;
-
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    @ColumnDefault("CURRENT_TIMESTAMP")
-    private LocalDateTime createdAt;
-}
