+++ backend/src/main/java/com/example/nbe341team02/domain/delivery/entity/Delivery.java
@@ -1,0 +1,41 @@
+package com.example.nbe341team02.domain.delivery.entity;
+
+import com.example.nbe341team02.domain.orders.entity.Order;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.CreationTimestamp;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+import java.util.HashSet;
+import java.util.Set;
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@EntityListeners(AuditingEntityListener.class)
+public class Delivery {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToMany(mappedBy = "delivery")
+    private Set<Order> orders = new HashSet<>();
+
+    @ManyToOne
+    private DeliveryCompany deliveryCompany;
+
+    @Setter
+    @Column
+    private String deliveryTrackingNumber;
+
+    @CreationTimestamp
+    @Column(nullable = false, updatable = false)
+    @ColumnDefault("CURRENT_TIMESTAMP")
+    private LocalDateTime createdAt;
+}
