+++ backend/src/main/java/com/example/nbe341team02/domain/delivery/repository/DeliveryCompanyRepository.java
@@ -1,5 +1,6 @@
-package com.example.nbe341team02.domain.delivery;
+package com.example.nbe341team02.domain.delivery.repository;
 
+import com.example.nbe341team02.domain.delivery.entity.DeliveryCompany;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Set;
