+++ backend/src/main/java/com/example/nbe341team02/domain/delivery/scheduler/DeliveryScheduler.java
@@ -1,5 +1,7 @@
-package com.example.nbe341team02.domain.delivery;
+package com.example.nbe341team02.domain.delivery.scheduler;
 
+import com.example.nbe341team02.domain.delivery.dto.DeliveryTimePolicyRegisterDto;
+import com.example.nbe341team02.domain.delivery.service.DeliveryService;
 import jakarta.annotation.PostConstruct;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
